Creation of final Master File
Creation of the final Master File with a write commit, where it processes and cleans the CSV file as downloaded from the three organizations (ADF, GEF, GCF), and joins them using 'project_no'. For AF we need to double check projects not saved and if ID that I used makes sense. 

The final block of the code consolidates the three different datasets based on 'project_no'. The schema of the Master file is as follows:
project_no	country	title	sector	entity	url	status	grant_amount